+++ src/main/java/com/dcr/api/model/dto/DcrproccStatusNew.java
@@ -1,5 +1,5 @@
 package com.dcr.api.model.dto;
 
-public record DcrproccStatusNew(Long idmatriz, String partnumpd, String tpprd, Integer status) {
+public record DcrproccStatusNew(Integer idmatriz, String partnumpd, String tpprd, Integer status) {
 
 }
