+++ src/main/java/com/dcr/api/model/dto/DcrprocchDTO.java
@@ -1,5 +1,5 @@
 package com.dcr.api.model.dto;
 
-public record DcrprocchDTO(Long idhist,Long idmatriz,String partnumpd,String tpprd,Integer stsold,Integer stsnew,String dtstatus,String hrstatus,String respstatus) {
+public record DcrprocchDTO(Long idhist,Integer idmatriz,String partnumpd,String tpprd,Integer stsold,Integer stsnew,String dtstatus,String hrstatus,String respstatus) {
 
 }
