+++ src/main/java/com/dcr/api/model/dto/DcrprotoDTO.java
@@ -1,5 +1,5 @@
 package com.dcr.api.model.dto;
 
-public record DcrprotoDTO(String protdcre,Long idmatriz,String partnumpd,String tpprd,String tpenvio,String dtenvio,String hrenvio,String repreenvio, String status) {
+public record DcrprotoDTO(String protdcre,Integer idmatriz,String partnumpd,String tpprd,String tpenvio,String dtenvio,String hrenvio,String repreenvio, String status) {
 
 }
